fix calculating recurring VAT - it is recalculated every time (in case Billing info changes)
