+++ plans/forms.py
@@ -1,36 +1,10 @@
 from django import forms
-from django.conf import settings
 from django.core.exceptions import ValidationError
 from django.forms.widgets import HiddenInput
 from django.utils.translation import ugettext
 
-from .models import PlanPricing, BillingInfo
-from plans.models import Order
-
-
-def get_client_ip(request):
-    x_forwarded_for = request.META.get('HTTP_X_FORWARDED_FOR')
-    if x_forwarded_for:
-        ip = x_forwarded_for.split(',')[0]
-    else:
-        ip = request.META.get('REMOTE_ADDR')
-    return ip
-
-
-def get_country_code(request):
-    if getattr(settings, 'PLANS_GET_COUNTRY_FROM_IP', False):
-        try:
-            from geolite2 import geolite2
-            reader = geolite2.reader()
-            ip_address = get_client_ip(request)
-            ip_info = reader.get(ip_address)
-        except ModuleNotFoundError:
-            ip_info = None
-
-        if ip_info and 'country' in ip_info:
-            country_code = ip_info['country']['iso_code']
-            return country_code
-    return getattr(settings, 'PLANS_DEFAULT_COUNTRY', None)
+from .models import BillingInfo, Order, PlanPricing
+from .utils import get_country_code
 
 
 class OrderForm(forms.Form):
